Full functionality complete.
Added mouseover events for the grid squares so they change according to the current
color selection. Can have colour via random hexcode or from rgb picker. Eraser button
included to remove unwanted sections & a clear button to wipe the slate.
Finally, there is also a slider to control the number of squares in the grid, which
is stepped to different sizes from 1x1 to 100x100.